requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0
tensorboard>=2.10.0
requests>=2.28.0

